# openapi.yaml
openapi: 3.0.0
info:
  title: Taska Backend API
  description: REST API для платформы Taska
  version: 1.0.0
  contact:
    name: Taska Support
    email: support@taska.com
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
    Task:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        column:
          type: string
        assignee:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        color:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        estimatedHours:
          type: number
          nullable: true
        totalHoursSpent:
          type: number
          nullable: true
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, archived]
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          nullable: true
...
paths:
...
  /api/projects:
    get:
      summary: Получить список проектов
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Создать новый проект
      tags:
        - Projects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  projectId:
                    type: string
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/{projectId}:
    get:
      summary: Получить информацию о проекте
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/projects/{projectId}/tasks:
    get:
      summary: Получить список задач проекта
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
        - in: query
          name: column
          schema:
            type: string
          description: Фильтр по колонке
        - in: query
          name: assignee
          schema:
            type: string
          description: Фильтр по исполнителю
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, done]
          description: Фильтр по статусу
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Нет доступа к проекту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Создать новую задачу
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - column
              properties:
                text:
                  type: string
                column:
                  type: string
                assignee:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum: [todo, in_progress, done]
                color:
                  type: string
                  nullable: true
                dueDate:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '201':
          description: Задача создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  taskId:
                    type: string
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Только PM может создавать задачи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...